Add tests for incremental compilation after a failure, also move all incremental compilation after a failure tests to a separate class
