+++ subprojects/language-java/src/integTest/groovy/org/gradle/java/compile/incremental/AbstractSourceIncrementalCompilationIntegrationTest.groovy
@@ -16,16 +16,10 @@
 
 package org.gradle.java.compile.incremental
 
-import org.gradle.api.internal.cache.StringInterner
-import org.gradle.api.internal.tasks.compile.incremental.recomp.PreviousCompilationAccess
 import org.gradle.integtests.fixtures.CompiledLanguage
 import spock.lang.Issue
 
-import static org.junit.Assume.assumeFalse
-
 abstract class AbstractSourceIncrementalCompilationIntegrationTest extends AbstractJavaGroovyIncrementalCompilationSupport {
-
-    abstract boolean isForkedCompiler();
 
     def "detects class changes in subsequent runs ensuring the class dependency data is refreshed"() {
         source "class A {}", "class B {}", "class C {}"
@@ -520,157 +514,5 @@ sourceSets {
         then:
         outputs.deletedClasses("Class\$Name")
         outputs.recompiledClasses("Class\$Name1", "Main")
-    }
-
-    def "old classes are are restored after the compile failure"() {
-        source("class A {}", "class B {}")
-        outputs.snapshot { run language.compileTaskName }
-
-        when:
-        source("class A { garbage }")
-        runAndFail language.compileTaskName
-
-        then:
-        outputs.noneRecompiled()
-        outputs.hasFiles(file("A.class"), file("B.class"))
-    }
-
-    def "incremental compilation works after a compile failure"() {
-        source("class A {}", "class B {}")
-        outputs.snapshot { run language.compileTaskName }
-        source("class A { garbage }")
-        runAndFail language.compileTaskName
-        outputs.noneRecompiled()
-
-        when:
-        source("class A { }")
-        run language.compileTaskName
-
-        then:
-        outputs.recompiledClasses("A")
-        outputs.hasFiles(file("A.class"), file("B.class"))
-    }
-
-    def "incremental compilation works after multiple compile failures"() {
-        source("class A {}", "class B {}")
-        outputs.snapshot { run language.compileTaskName }
-        source("class A { garbage }")
-        runAndFail language.compileTaskName
-        outputs.noneRecompiled()
-        runAndFail language.compileTaskName
-        outputs.noneRecompiled()
-
-        when:
-        source("class A { }")
-        run language.compileTaskName
-
-        then:
-        outputs.recompiledClasses("A")
-        outputs.hasFiles(file("A.class"), file("B.class"))
-    }
-
-    def "nothing is recompiled after a compile failure when file is reverted"() {
-        source("class A {}", "class B {}")
-        outputs.snapshot { run language.compileTaskName }
-        source("class A { garbage }")
-        runAndFail language.compileTaskName
-        outputs.noneRecompiled()
-
-        when:
-        source("class A {}")
-        run language.compileTaskName
-
-        then:
-        outputs.noneRecompiled()
-        outputs.hasFiles(file("A.class"), file("B.class"))
-    }
-
-    def "writes relative paths to source-to-class mapping after incremental compilation"() {
-        given:
-        assumeFalse("Source-to-class mapping is not written for CLI compiler", this.class == JavaSourceCliIncrementalCompilationIntegrationTest.class)
-        source("package test.gradle; class A {}", "package test2.gradle; class B {}")
-        outputs.snapshot { run language.compileTaskName }
-        def previousCompilationDataFile = file("build/tmp/${language.compileTaskName}/previous-compilation-data.bin")
-        def previousCompilationAccess = new PreviousCompilationAccess(new StringInterner())
-
-        when:
-        source("package test.gradle; class A { static class C {} }")
-        run language.compileTaskName
-
-        then:
-        outputs.recompiledClasses("A", 'A$C')
-        def previousCompilationData = previousCompilationAccess.readPreviousCompilationData(previousCompilationDataFile)
-        previousCompilationData.compilerApiData.sourceToClassMapping == [
-            ("test/gradle/A.${languageName}" as String): ["test.gradle.A", "test.gradle.A\$C"] as Set,
-            ("test2/gradle/B.${languageName}" as String): ["test2.gradle.B"] as Set
-        ]
-    }
-
-    def "nothing is stashed on full recompilation but it's stashed on incremental compilation"() {
-        given:
-        source("class A {}", "class B {}")
-        def compileTransactionDir = file("build/tmp/${language.compileTaskName}/compileTransaction")
-
-        when: "First compilation is always full compilation"
-        run language.compileTaskName
-
-        then:
-        outputs.recompiledClasses("A", "B")
-        !compileTransactionDir.exists()
-
-        when: "Compilation after failure with clean is full recompilation"
-        source("class A { garbage }")
-        runAndFail language.compileTaskName
-        outputs.snapshot { source("class A { }") }
-        run "clean", language.compileTaskName
-
-        then:
-        outputs.recompiledClasses("A", "B")
-        !compileTransactionDir.exists()
-
-        when: "Incremental compilation"
-        outputs.snapshot { source("class A {}") }
-        run language.compileTaskName
-
-        then:
-        outputs.recompiledClasses("A")
-        compileTransactionDir.exists()
-        compileTransactionDir.file("stash-dir/A.class.uniqueId0").exists()
-    }
-
-    def "unrelated class in the same source file as affected class is recompiled and stashed on incremental compilation"() {
-        given:
-        // For forked compiler we don't have a source-class-mapping,so we can't detect such class
-        assumeFalse(isForkedCompiler())
-        source "class A {}", "class B extends A {}", "class D {}"
-        // F is not affected by changes, but it is recompiled
-        // and still stashed since it's in the same source file as C
-        source """
-        class C extends B {}
-        class F {}
-        """
-        def compileTransactionDir = file("build/tmp/${language.compileTaskName}/compileTransaction")
-        def stashDirClasses = {
-            compileTransactionDir.file("stash-dir")
-                .list()
-                .collect { it.replaceAll(".uniqueId.", "") }
-                .sort()
-        }
-
-        when: "First compilation is always full compilation"
-        run language.compileTaskName
-
-        then:
-        outputs.recompiledClasses("A", "B", "C", "D", "F")
-        !compileTransactionDir.exists()
-
-        when:
-        outputs.snapshot { source("class A { /* comment */ }") }
-        run language.compileTaskName
-
-        then:
-        outputs.recompiledClasses("A", "B", "C", "F")
-        compileTransactionDir.exists()
-        stashDirClasses() == ["A.class", "B.class", "C.class", "F.class"]
     }
 }
