+++ subprojects/language-java/src/integTest/groovy/org/gradle/java/compile/incremental/JavaSourceCliIncrementalCompilationIntegrationTest.groovy
@@ -30,9 +30,4 @@ class JavaSourceCliIncrementalCompilationIntegrationTest extends BaseJavaSourceI
             }
         """
     }
-
-    @Override
-    boolean isForkedCompiler() {
-        return true
-    }
 }
