+++ subprojects/language-java/src/integTest/groovy/org/gradle/java/compile/incremental/JavaSourceIncrementalCompilationIntegrationTest.groovy
@@ -181,9 +181,4 @@ class JavaSourceIncrementalCompilationIntegrationTest extends BaseJavaSourceIncr
         then:
         outputs.recompiledClasses("A", "B", "E", "Const", "package-info")
     }
-
-    @Override
-    boolean isForkedCompiler() {
-        return false
-    }
 }
