+++ subprojects/language-java/src/main/java/org/gradle/api/internal/tasks/compile/incremental/recomp/PreviousCompilation.java
@@ -17,11 +17,9 @@
 package org.gradle.api.internal.tasks.compile.incremental.recomp;
 
 import org.gradle.api.internal.tasks.compile.incremental.compilerapi.deps.DependentsSet;
-import org.gradle.api.internal.tasks.compile.incremental.compilerapi.deps.GeneratedResource;
 import org.gradle.api.internal.tasks.compile.incremental.deps.ClassSetAnalysis;
 
 import javax.annotation.Nullable;
-import java.util.Collections;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
